+++ site/astro.config.mjs
@@ -1,12 +1,11 @@
 import { defineConfig } from "astro/config";
 import react from "@astrojs/react";
-import tailwind from "@astrojs/tailwind";
 import sitemap from "@astrojs/sitemap";
 
 export default defineConfig({
-  site: "https://sachncs.github.io",
+  site: "https://sachncs.github.io/cedrus",
   base: "/cedrus",
-  integrations: [react(), tailwind({ applyBaseStyles: false }), sitemap()],
+  integrations: [react(), sitemap()],
   build: {
     assets: "assets",
   },